requirements.txt
flask==2.0.1
requests==2.26.0
requests-oauthlib==1.3.0
python-dotenv==0.19.0
Werkzeug==2.0.0
gunicorn
